A crash when deleting an item is fixed now. Check picturePath to see if it is null.

+++ app/src/main/java/yiwejeje/staticrecallapp/Model/Item.java
@@ -125,8 +125,11 @@ public class Item implements Serializable {
     }
 
     public void deleteImage() {
-        File imageFile = new File(picturePath);
-        imageFile.delete();
+        if (picturePath != null) {
+            File imageFile = new File(picturePath);
+            imageFile.delete();
+        }
+
     }
 
     public String toString() {
